requirements.txt
black
fastapi
pytest
pylint
uvicorn
sqlalchemy
datetime
uuid
shortuuid
psycopg2-binary
python-multipart
boto3
python-dotenv
tk
pandas
numpy
awswrangler
fastapi-crudrouter
passlib
bcrypt
shortuuid
cffi
click
h11
pycparser
pydantic
PyJWT
six
starlette
uvicorn
scipy
matplotlib
scikit-learn 
IPython
python-slugify